requirements.txt
Django>=1.10,<1.11
Pillow>=3.1.1
django-crispy-forms>=1.6.1
django-extensions>=1.7.6
django-filter>=1.0
django-flat-responsive>=1.2.0
djangorestframework>=3.5.3
filemagic>=1.6
fuzzywuzzy[speedup]==0.15.0
langdetect>=1.0.7
https://github.com/openpaperwork/pyocr/zipball/30182d153535489588fec630f0c533f49334e4b2#egg=pyocr
python-dateutil>=2.6.0
python-dotenv>=0.6.2
python-gnupg>=0.3.9
pytz>=2016.10
gunicorn==19.7.1
chardet==3.0.4
PyPDF2==1.26.0

# For the tests
factory-boy
pytest
pytest-django
pytest-sugar
pytest-env
pep8
flake8
tox
